Update a basic statistics template
