+++ app/src/main/java/com/gasperpintar/smokingtracker/ui/fragment/statistics/BasicFragment.kt
@@ -1,0 +1,28 @@
+package com.gasperpintar.smokingtracker.ui.fragment.statistics
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.gasperpintar.smokingtracker.databinding.FragmentStatisticsBasicBinding
+
+class BasicFragment : Fragment() {
+
+    private var _binding: FragmentStatisticsBasicBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentStatisticsBasicBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
