+++ app/src/main/java/com/gasperpintar/smokingtracker/ui/fragment/statistics/ForecastFragment.kt
@@ -1,0 +1,27 @@
+package com.gasperpintar.smokingtracker.ui.fragment.statistics
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.gasperpintar.smokingtracker.databinding.FragmentStatisticsForecastBinding
+
+class ForecastFragment : Fragment() {
+    private var _binding: FragmentStatisticsForecastBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentStatisticsForecastBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
